T#800 P3-F: Extract queue/gorn module from server.ts
Sub-task F of T#786 Phase 3. Extract 3 queue/gorn handlers (~70 LoC) from
server.ts into src/queue/routes.ts.

## Changes

- New: src/queue/routes.ts (83L) — registerQueueRoutes(app, sqlite, helpers)
  with QueueHelpers DI (isTrustedRequest), 3 handlers extracted verbatim
- server.ts: 4228L → 4160L (-68L)

## Handlers
- GET /api/queue/gorn — list queue items by status
- POST /api/queue/gorn — add thread to gorn-queue (any Beast can tag)
- PATCH /api/queue/gorn/:threadId — update queue status (decided/deferred/pending/withdrawn, gorn-only browser gate via isTrustedRequest fallback)

## Test plan
- [x] bun build clean (Norm #84 Layer 1)
- [x] Module registration before SPA fallback (T#783 lesson)
- [ ] Post-deploy smoke at deploy gate

Reviewer: @zaghnal per task assignment.
